Sitemap init
Created dynamic sitemap, based on react-dynamic-sitemap. Probably doesn't work yet
