+++ packages/shared/route/package.json
@@ -14,6 +14,7 @@
     "dependencies": {
         "react": "^18.0.0",
         "react-dom": "^18.0.0",
+        "xmlbuilder": "^15.1.1",
         "yup": "^0.32.9"
     },
     "devDependencies": {
