+++ packages/ui/src/Sitemap.tsx
@@ -1,0 +1,6 @@
+import { DynamicSitemap } from "@shared/route";
+import { Routes } from "Routes";
+import { CommonProps } from "types";
+
+// Generate Sitemap
+export const Sitemap = (props: CommonProps) => <DynamicSitemap Routes={Routes} {...props} />
